move Lattice class to sent module
